# servers/cromwell/jobs/models/event.py
# coding: utf-8

from __future__ import absolute_import
from .base_model_ import Model
from datetime import date, datetime
from typing import List, Dict
from ..util import deserialize_model


class Event(Model):
    """
    NOTE: This class is auto generated by the swagger code generator program.
    Do not edit the class manually.
    """
    def __init__(self, time=None, content=None):
        """
        Event - a model defined in Swagger

        :param time: The time of this Event.
        :type time: str
        :param content: The content of this Event.
        :type content: str
        """
        self.swagger_types = {
            'time': str,
            'content': str
        }

        self.attribute_map = {
            'time': 'time',
            'content': 'content'
        }

        self._time = time
        self._content = content

    @classmethod
    def from_dict(cls, dikt):
        """
        Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Event of this Event.
        :rtype: Event
        """
        return deserialize_model(dikt, cls)

    @property
    def time(self):
        """
        Gets the time of this Event.

        :return: The time of this Event.
        :rtype: str
        """
        return self._time

    @time.setter
    def time(self, time):
        """
        Sets the time of this Event.

        :param time: The time of this Event.
        :type time: str
        """

        self._time = time

    @property
    def content(self):
        """
        Gets the content of this Event.

        :return: The content of this Event.
        :rtype: str
        """
        return self._content

    @content.setter
    def content(self, content):
        """
        Sets the content of this Event.

        :param content: The content of this Event.
        :type content: str
        """

        self._content = content